refactor(Sales): Move brand palette helpers into branding, tidy workflow
- Move the brand colour-palette helpers (_to_hex, accent_ramp,
  quantile_palette, accent_by_magnitude) out of al_utils (which is for
  AlphaLens helpers) and into branding.py, where the rest of the brand
  presentation lives; update page_one/page_two imports accordingly.
- generate-factsheets.yml: drop the stale top-of-file delivery preamble
  and the "delivery"/"fast iteration"/"steady state" inline comments;
  correct the font comment (Mona Sans is the bundled brand font; Inter
  is only a defensive fallback).
- Refresh page_one's module docstring: it renders performance/risk
  tables now, not the monthly heatmap.
